Fix a tolerance-overload bug that made intercept precision asserts inert
`Precision` was declared `const float`, so `Assert.Equal(a, b, Precision)`
bound to xunit's absolute-tolerance overload (allowing any two values
within 4.0 of each other) instead of the intended 4-decimal-place
overload. Hand-mutating the `t2 > 0 ? min(t1, t2) : t1` ternary to always
return `t1` (math.cs:287) exposed it: the assertion silently accepted a
result almost 0.27 away from the expected value. Declaring it `const int`
selects the correct overload; confirmed the same hand mutation now fails.

Also add the float2-overload euler-integration length-validation test the
mutation report showed missing (BatchMath.cs:208's ValidateEqualLengths
call had no test at all).

+++ packages/cultmath/tests/CultMath.Tests/FirstOrderInterceptTests.cs
@@ -24,7 +24,7 @@ namespace CultMath.Tests;
 // determinant > 0 with both roots <= 0 => 0 (via max(t2, 0)).
 public sealed class FirstOrderInterceptTests
 {
-    private const float Precision = 4;
+    private const int Precision = 4;
 
     [Fact]
     public void ZeroRelativeVelocityReturnsZeroTime()
